--- a/lib/collection/src/grouping/aggregator.rs
+++ b/lib/collection/src/grouping/aggregator.rs
@@ -1,5 +1,5 @@
 use std::collections::hash_map::Entry;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 
 use itertools::Itertools;
 use ordered_float::OrderedFloat;
@@ -9,31 +9,34 @@ use segment::types::{
     ExtendedPointId, Order, PayloadContainer, PointIdType, ScoreType, ScoredPoint,
 };
 use serde_json::Value;
+use ahash::{AHashMap, AHashSet};
 
 use super::types::AggregatorError::{self, *};
 use super::types::Group;
 
-type Hits = HashMap<PointIdType, ScoredPoint>;
+type Hits = AHashMap<PointIdType, ScoredPoint>;
 
 pub(super) struct GroupsAggregator {
-    groups: HashMap<GroupId, Hits>,
+    groups: AHashMap<GroupId, Hits>,
     max_group_size: usize,
     grouped_by: String,
     max_groups: usize,
-    full_groups: HashSet<GroupId>,
-    group_best_scores: HashMap<GroupId, ScoreType>,
+    full_groups: AHashSet<GroupId>,
+    group_best_scores: AHashMap<GroupId, ScoreType>,
+    all_ids: HashSet<ExtendedPointId>,
     order: Order,
 }
 
 impl GroupsAggregator {
     pub(super) fn new(groups: usize, group_size: usize, grouped_by: String, order: Order) -> Self {
         Self {
-            groups: HashMap::with_capacity(groups),
+            groups: AHashMap::with_capacity(groups),
             max_group_size: group_size,
             grouped_by,
             max_groups: groups,
-            full_groups: HashSet::with_capacity(groups),
-            group_best_scores: HashMap::with_capacity(groups),
+            full_groups: AHashSet::with_capacity(groups),
+            group_best_scores: AHashMap::with_capacity(groups),
+            all_ids: HashSet::with_capacity(groups * group_size),
             order,
         }
     }
@@ -68,7 +71,7 @@ impl GroupsAggregator {
             let group = self
                 .groups
                 .entry(group_key.clone())
-                .or_insert_with(|| HashMap::with_capacity(self.max_group_size));
+                .or_insert_with(|| AHashMap::with_capacity(self.max_group_size));
 
             let entry = group.entry(point.id);
 
@@ -81,6 +84,7 @@ impl GroupsAggregator {
                 }
                 Entry::Vacant(v) => {
                     v.insert(point.clone());
+                    self.all_ids.insert(point.id);
                 }
             }
 
@@ -130,7 +134,7 @@ impl GroupsAggregator {
 
     /// Gets the keys of the groups that have less than the max group size
     pub(super) fn keys_of_unfilled_best_groups(&self) -> Vec<Value> {
-        let best_group_keys: HashSet<_> = self.best_group_keys().cloned().collect();
+        let best_group_keys: AHashSet<_> = self.best_group_keys().cloned().collect();
         best_group_keys
             .difference(&self.full_groups)
             .cloned()
@@ -145,17 +149,13 @@ impl GroupsAggregator {
 
     /// Gets the amount of best groups that have reached the max group size
     pub(super) fn len_of_filled_best_groups(&self) -> usize {
-        let best_group_keys: HashSet<_> = self.best_group_keys().cloned().collect();
+        let best_group_keys: AHashSet<_> = self.best_group_keys().cloned().collect();
         best_group_keys.intersection(&self.full_groups).count()
     }
 
     /// Gets the ids of the already present points across all the groups
-    pub(super) fn ids(&self) -> HashSet<ExtendedPointId> {
-        self.groups
-            .iter()
-            .flat_map(|(_, hits)| hits.iter())
-            .map(|(pid, _)| *pid)
-            .collect()
+    pub(super) fn ids(&self) -> &HashSet<ExtendedPointId> {
+        &self.all_ids
     }
 
     /// Returns the best groups sorted by their best hit. The hits are sorted too.