--- a/lib/collection/src/grouping/group_by.rs
+++ b/lib/collection/src/grouping/group_by.rs
@@ -284,12 +284,8 @@ where
     );
 
     // Try to complete amount of groups
+    let mut needs_filling = true;
     for _ in 0..MAX_GET_GROUPS_REQUESTS {
-        // TODO: should we break early if we have some amount of "enough" groups?
-        if aggregator.len_of_filled_best_groups() >= request.limit {
-            break;
-        }
-
         let mut request = request.clone();
 
         let source = &mut request.source;
@@ -308,51 +304,7 @@ where
         }
 
         // exclude already aggregated points
-        let ids = aggregator.ids();
-        if !ids.is_empty() {
-            let exclude_ids = Filter::new_must_not(Condition::HasId(ids.into()));
-            source.merge_filter(&exclude_ids);
-        }
-
-        let points = request
-            .r#do(
-                collection,
-                collection_by_name.clone(),
-                read_consistency,
-                shard_selection,
-            )
-            .await?;
-
-        if points.is_empty() {
-            break;
-        }
-
-        aggregator.add_points(&points)
-    }
-
-    // Try to fill up groups
-    for _ in 0..MAX_GROUP_FILLING_REQUESTS {
-        if aggregator.len_of_filled_best_groups() >= request.limit {
-            break;
-        }
-
-        let mut request = request.clone();
-
-        let source = &mut request.source;
-
-        // construct filter to only include unsatisfied groups
-        let unsatisfied_groups = aggregator.keys_of_unfilled_best_groups();
-        let match_any = match_on(&request.group_by, unsatisfied_groups);
-        if !match_any.is_empty() {
-            let include_groups = Filter {
-                must: Some(match_any),
-                ..Default::default()
-            };
-            source.merge_filter(&include_groups);
-        }
-
-        // exclude already aggregated points
-        let ids = aggregator.ids();
+        let ids = aggregator.ids().clone();
         if !ids.is_empty() {
             let exclude_ids = Filter::new_must_not(Condition::HasId(ids.into()));
             source.merge_filter(&exclude_ids);
@@ -372,6 +324,58 @@ where
         }
 
         aggregator.add_points(&points);
+
+        // TODO: should we break early if we have some amount of "enough" groups?
+        if aggregator.len_of_filled_best_groups() >= request.limit {
+            needs_filling = false;
+            break;
+        }
+    }
+
+    // Try to fill up groups
+    if needs_filling {
+        for _ in 0..MAX_GROUP_FILLING_REQUESTS {
+            let mut request = request.clone();
+
+            let source = &mut request.source;
+
+            // construct filter to only include unsatisfied groups
+            let unsatisfied_groups = aggregator.keys_of_unfilled_best_groups();
+            let match_any = match_on(&request.group_by, unsatisfied_groups);
+            if !match_any.is_empty() {
+                let include_groups = Filter {
+                    must: Some(match_any),
+                    ..Default::default()
+                };
+                source.merge_filter(&include_groups);
+            }
+
+            // exclude already aggregated points
+            let ids = aggregator.ids().clone();
+            if !ids.is_empty() {
+                let exclude_ids = Filter::new_must_not(Condition::HasId(ids.into()));
+                source.merge_filter(&exclude_ids);
+            }
+
+            let points = request
+                .r#do(
+                    collection,
+                    collection_by_name.clone(),
+                    read_consistency,
+                    shard_selection,
+                )
+                .await?;
+
+            if points.is_empty() {
+                break;
+            }
+
+            aggregator.add_points(&points);
+            
+            if aggregator.len_of_filled_best_groups() >= request.limit {
+                break;
+            }
+        }
     }
 
     // extract best results