--- a/lib/collection/src/grouping/group_by.rs
+++ b/lib/collection/src/grouping/group_by.rs
@@ -371,7 +371,7 @@ where
             }
 
             aggregator.add_points(&points);
-            
+
             if aggregator.len_of_filled_best_groups() >= request.limit {
                 break;
             }