--- a/lib/segment/src/index/field_index/full_text_index/text_index.rs
+++ b/lib/segment/src/index/field_index/full_text_index/text_index.rs
@@ -123,6 +123,10 @@ impl FullTextIndex {
         self.get_doc(point_id).map(|x| x.len()).unwrap_or(0)
     }
 
+    pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
+        self.get_doc(point_id).map(|x| x.is_empty()).unwrap_or(true)
+    }
+
     pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
         self.get_doc(point_id).is_none()
     }