--- a/lib/segment/src/index/field_index/full_text_index/text_index.rs
+++ b/lib/segment/src/index/field_index/full_text_index/text_index.rs
@@ -122,6 +122,10 @@ impl FullTextIndex {
         // Maybe we want number of documents in the future?
         self.get_doc(point_id).map(|x| x.len()).unwrap_or(0)
     }
+
+    pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
+        self.get_doc(point_id).is_none()
+    }
 }
 
 impl ValueIndexer<String> for FullTextIndex {