--- a/lib/segment/src/index/field_index/geo_index.rs
+++ b/lib/segment/src/index/field_index/geo_index.rs
@@ -469,6 +469,10 @@ impl GeoMapIndex {
     pub fn values_count(&self, point_id: PointOffsetType) -> usize {
         self.get_values(point_id).map(|x| x.len()).unwrap_or(0)
     }
+    
+    pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
+        self.get_values(point_id).is_none()
+    }
 }
 
 impl ValueIndexer<GeoPoint> for GeoMapIndex {