--- a/lib/segment/src/index/field_index/geo_index.rs
+++ b/lib/segment/src/index/field_index/geo_index.rs
@@ -471,9 +471,11 @@ impl GeoMapIndex {
     }
 
     pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
-        self.get_values(point_id).map(|x| x.is_empty()).unwrap_or(true)
+        self.get_values(point_id)
+            .map(|x| x.is_empty())
+            .unwrap_or(true)
     }
-    
+
     pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
         self.get_values(point_id).is_none()
     }