--- a/lib/segment/src/index/field_index/map_index.rs
+++ b/lib/segment/src/index/field_index/map_index.rs
@@ -184,7 +184,9 @@ impl<N: Hash + Eq + Clone + Display + FromStr> MapIndex<N> {
     }
 
     pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
-        self.get_values(point_id).map(|x| x.is_empty()).unwrap_or(true)
+        self.get_values(point_id)
+            .map(|x| x.is_empty())
+            .unwrap_or(true)
     }
 
     pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {