--- a/lib/segment/src/index/field_index/numeric_index.rs
+++ b/lib/segment/src/index/field_index/numeric_index.rs
@@ -298,9 +298,14 @@ impl<T: Encodable + Numericable> NumericIndex<T> {
         self.get_values(point_id).map(|x| x.len()).unwrap_or(0)
     }
 
+    pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
+        self.get_values(point_id).map(|x| x.is_empty()).unwrap_or(true)
+    }
+
     pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
         self.get_values(point_id).is_none()
     }
+    
 }
 
 impl<T: Encodable + Numericable> PayloadFieldIndex for NumericIndex<T> {