--- a/lib/segment/src/index/field_index/numeric_index.rs
+++ b/lib/segment/src/index/field_index/numeric_index.rs
@@ -299,13 +299,14 @@ impl<T: Encodable + Numericable> NumericIndex<T> {
     }
 
     pub fn values_is_empty(&self, point_id: PointOffsetType) -> bool {
-        self.get_values(point_id).map(|x| x.is_empty()).unwrap_or(true)
+        self.get_values(point_id)
+            .map(|x| x.is_empty())
+            .unwrap_or(true)
     }
 
     pub fn values_is_none(&self, point_id: PointOffsetType) -> bool {
         self.get_values(point_id).is_none()
     }
-    
 }
 
 impl<T: Encodable + Numericable> PayloadFieldIndex for NumericIndex<T> {