Adds a handful of optimizations to make group_by faster. Measured against `dev` using a new internal benchmark: `lib/collection/benches/search_groups.rs`

## Overall Improvements
- [x] Make `is_empty` take advantage of the index when not empty.
- [x] Make `is_null` take advantage of the index when not null

## `group_by` Improvements
- [x] improve termination condition check -> mostly just refactor
- [x] add all_ids to aggregator to avoid flat_map -> found in flamegraph
- [x] use `ahash` for aggregator hashmaps and hashsets (debatable)

These are the perf changes so far:

| scenario/parameters | time | improvement |
|--------|--------|--------|
| unindexed-search-groups/group_size=1, limit=1 | 1.2771 ms | No change detected |
| unindexed-search-groups/group_size=1, limit=10 |  1.6891 ms | +27.076% |
| unindexed-search-groups/group_size=1, limit=100 | 5.0947 ms | +13.410% |
| unindexed-search-groups/group_size=1, limit=1000 | 202.43 m | No change detected |
| unindexed-search-groups/group_size=10, limit=1 | 71.322 ms | No change detected | 
| unindexed-search-groups/group_size=10, limit=10 | 133.02 ms | No change detected |
| unindexed-search-groups/group_size=10, limit=100 | 50.543 ms | -10.749% | 
| unindexed-search-groups/group_size=10, limit=1000 | 218.36 ms | -5.7298% | 
| unindexed-search-groups/group_size=100, limit=1 | 80.092 ms | +4.7103% | 
| unindexed-search-groups/group_size=100, limit=10 | 156.51 ms | Change within noise threshold | 
| unindexed-search-groups/group_size=100, limit=100 | 173.09 ms | -8.5538% |
| unindexed-search-groups/group_size=100, limit=1000 | 419.80 ms | -33.124% | 
| indexed-search-groups/group_size=1, limit=1 | 540.66 µs | -33.124% | 
| indexed-search-groups/group_size=1, limit=10 | 760.93 µs | -47.628% |
| indexed-search-groups/group_size=1, limit=100 |  2.8926 ms | -29.044% | 
| indexed-search-groups/group_size=1, limit=1000 | 107.52 ms | -12.617% |
| indexed-search-groups/group_size=10, limit=1 | 16.421 ms | -17.733%|
| indexed-search-groups/group_size=10, limit=10 | 35.713 ms | -18.842% | 
| indexed-search-groups/group_size=10, limit=100 | 35.928 ms | -27.525% |
| indexed-search-groups/group_size=10, limit=1000 | 147.72 ms | -17.593% |
| indexed-search-groups/group_size=100, limit=1 | 20.611 ms | -17.175% | 
| indexed-search-groups/group_size=100, limit=10 | 63.723 ms | -15.518% |
| indexed-search-groups/group_size=100, limit=100 | 165.04 ms | -6.0571% |
| indexed-search-groups/group_size=100, limit=1000 | 369.62 ms | -22.634% | 

### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why you'd like us to include them?
* [x] Have you successfully ran tests with your changes locally?